Added Spectrogram Visualization of FFT
To see the spectrogram value over time, I have ditched the line-based visualization and created a spectrogram visualization.

Each frame, the spectrum values will be used to display a single bar of colors, with the color and brightness indicating the intensity of each band. This bar moves along to the right over time, allowing you to see the spectrogram graph.