Added Ability to Record Beat Indices
To train the model on what counts as a beat, we must create a dataset that contains snippets of FFT data and whether or not it is a beat. To start, I have implemented a way to manually label parts of the song as beats using the mouse.

In this commit, I do the following:
- Song data is now saved when the recording ends and loaded at the start of the program. That way you don't have to record the same song's values every time you run the program.
- Implemented scrolling so that the user can navigate the values of the song wave.
- Left clicking the mouse adds the index of the value underneath the mouse to the list of beat indices and saves their data.
- Right clicking the mouse removes the index of the value the mouse and saves the data.