Added Cursor to Playback the Song
Previously, once the song had finished recording, it wouldn't be played again.

Now, I have added a cursor that marks the start of the playback of the song. Pressing the keyboard now toggles playing and pausing the song. The middle mouse button is used to set the position of the cursor.

Also, a small circle visualizes when a beat is detected to help sync the beats to the song.