Added Multiple Songs Support
The program now allows you to load several song files inside a new "Songs" folder so that you can record the data for each file all in one session. Simply put any audio file with a .mp3 or .wav extension inside the "Songs" folder.

The program will go through the songs one at a time. When you finish selecting the beat indices of the current song, you can press the 'N' key to take you to the next song.

Whenever the program detects that you haven't recorded a song's values, it will enter recording mode and you have to wait for the recording to finish.

Lastly, all average values are normalized between 0 and 1 to remain consistent for the model.